+++ CHANGELOG.md
@@ -6,11 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add email notification support for newly created users via csv import (#125)
 - Add Data Synchronisation Detail Component to notify administrative users about cas / dogu behaviour (#119)
 
 ### Changed
 - Update ces-theme-tailwind to v0.3.8
 
+### Fixed
+- Fix column overflow in user import table
 
 ## [v1.13.0-1] - 2024-02-29
 ### Added